*: revert time back to u64 as it's a bit silly

+++ backend/crates/eiffelvis_core/src/domain/user_queries.rs
@@ -23,8 +23,8 @@ pub enum Filter {
     None,
     /// Open ended range of time
     Time {
-        begin: Option<u128>,
-        end: Option<u128>,
+        begin: Option<u64>,
+        end: Option<u64>,
     },
     /// Event Type
     Type { name: String },
